[sdk]: rename query-client.ts -> queryClient.ts
