+++ sdk/packages/sdk/src/protocols/intents/IntentGateway.ts
@@ -16,7 +16,7 @@ import type { ResumeIntentOrderOptions } from "@/types"
 import type { IEvmChain } from "@/chain"
 import type { IntentsCoprocessor } from "@/chains/intentsCoprocessor"
 import type { IsmpClient } from "@/client"
-import { _queryOrderInternal } from "@/query-client"
+import { _queryOrderInternal } from "@/queryClient"
 import type { IntentGatewayContext } from "./types"
 import type { CancelEvent } from "./types"
 import { CryptoUtils } from "./CryptoUtils"
