+++ sdk/packages/sdk/src/tests/tokenGateway.test.ts
@@ -22,7 +22,7 @@ import { hexToU8a, u8aToHex } from "@polkadot/util"
 import type { KeyringPair } from "@polkadot/keyring/types"
 import { encodeISMPMessage, EvmChain, SubstrateChain } from "@/chain"
 import { __test, ADDRESS_ZERO, bytes20ToBytes32 } from "@/utils"
-import { createQueryClient } from "@/query-client"
+import { createQueryClient } from "@/queryClient"
 import { IsmpClient } from "@/client"
 import { privateKeyToAccount, privateKeyToAddress } from "viem/accounts"
 import { baseSepolia, bscTestnet } from "viem/chains"
